Sending eMails to the given parameter partly working
different parameter and parameter types can now be handled due to the three ifs at the top and the seperation between role, ID and email at the bottom. emails already work fine. userIDs maybe work but i dont know, query is hard to debug. roles works so far but it doesnt filter the results to restrict them to a schoolID right now.