fixed the userIds and promise now works
$schoolId din query doesnt work and causes errors
userId can now be given as string, as Object Id, as array of strings and as array of objectIds. no userid at all also works.
promise works, if there are only email adresses the promise wont trigger (cause there are no promises, obviously) so the if-else is for that case.
Last step: get the query for roles working with restriction to school