Sc 170 email benachrichtigung für admin (#310)
* handcode smtp string due to problems

handcoded smtp string because secrets smtp dont return the windows set SMTP variable

* new hook sendMail

created the hook to send the eMail with, calling that hook as an afterhook from helpdesk. currently in get for testing

* added some text to the email

Some text is in the email now. schould send at post. i dont know if it works

* new global allmighty mail function

global.sendEmail with lots of parameters for sending mails server-side to replace all existing most client-side mailings
lots of stuff to do

* sendEmail now gets an infotext

I made the var infoText into a function that creates the infoText when given the missing parameters. that function is called in the afterhooks create and returns a correct infotext. that infotext is then given to sendEmail.

* make mails great again!

rebuild mailing so that it works again, some todos left for katrin

* small adjustments

small adjustments

* Sending eMails to the given parameter partly working

different parameter and parameter types can now be handled due to the three ifs at the top and the seperation between role, ID and email at the bottom. emails already work fine. userIDs maybe work but i dont know, query is hard to debug. roles works so far but it doesnt filter the results to restrict them to a schoolID right now.

* prepare for katrin

little changes and more todos

* receipients array

the email receipients are all written into one array (using promises for the db stuff) and then the emails are send in the end, checking the array for duplicates.

* Added Promise and eMail RegEx

Added Promise, doesnt work always bc sometime there are multiple POST requests and the promise breaks.

erased not needed lines of code

added reg ex to check if email adress is valide. very general regex

* fixed the userIds and promise now works

$schoolId din query doesnt work and causes errors
userId can now be given as string, as Object Id, as array of strings and as array of objectIds. no userid at all also works.
promise works, if there are only email adresses the promise wont trigger (cause there are no promises, obviously) so the if-else is for that case.
Last step: get the query for roles working with restriction to school

* query restriction to schoolId now works

* erased comments

* beautify code

not sure if working bc i cant test it bc everything sucks

* erased changes to package.json, erased debug changes in service.js, added .vscode to gitignore

* userIds cannot be given as ObjectId

* removed vscode launch.json

* lodash, other small refactorings

* var->let

* Fixed some issues from PR

* added error handling

* better solution for the two cases

* let receipientsDuplicatefree not needet

* try catch around mail to .catch, added limit to find-request, dont know what limit makes sense

* fix spacesVsTabs for travis

* modify helpdesk tests

travis didnt like hook.params.account.username, so work around