SC-9213 use getAllSubmissionsByUser in taskrepo via DIP
taskRepo.findAllOpenByStudent needs to access all submissions of a user. passing this via parameter would make the function hard to use, since its primary objective is to filter tasks that have submissions (so only open tasks are left). On the other hand, we should avoid repositories knowing other repositories.

moving the logic into UC layer would be recommended, but requires larger refactorings. In the meantime, it was decided that since submissions are a concept based on tasks, it would be okay for the submission repo to know about the task repo. In order for the task repo to access a submission repo, the dependency inversion principle has been applied, defining an interface in the task repo that the submission repo now implements.