extract dto dependency from task UC (#2846)
Both the student and the teacher usecase are now returning the same type, containing task entities and metadata about the submission status. This type is also returned by the combined usecase. This removes the dependency on the dto (that belongs to the controller)

The function calculating the submission status has been moved into the domain layer, into a domain service.