detect cycles in role inheritance structure
Assuming the former role structure is without cycles, a change in the role structure can be detected by examining the children of a newly changed role.
Each children, and recursively all childrens children, have to be checked if they equal the role where the change occured.
In case of a match, a cycle has been detected and an error has to be thrown. If no match is found, the inheritance tree of the roles is still cycle-free.
This essentially is a very simple partial BFS starting on the changed role.