BC-247 - remove timestamp from LDAP search Query for sync (#3010)
After the change to an other more parallel sync system. In timestamp is not more necessary for a delta to the last sync so we sync all data at once.
And from time to time the method to add the time stamp work not so well and then the ldap search query where wrong, but work.