refactor gridElement Entity to not contain positions
In the old implementation, changing a position would have involved getting a gridElement, and asking it to change its position.
However, the element would not know about the positions of other elements, hence it would not be able to confirm the validity of this request.
To avoid this issue, the positions are only known to the dashboard itself, which can make changes accordingly.

For easier access within the entity, it stores the elements now in a map. This is hidden to the outside, and both the constructor and the getter work with arrays still.

However, the way these arrays look have changed, with the position being seperated from the rest of the data. dto mapper and repo mapper have changed accordingly.

The data model remains the same, with position being saved in the GridElementModelEntity.