refactor gridelement to be created without id
For ungrouping elements we need to be able to create new gridelements
from the dashboard entity. Since it does not know about the repo, nor
the id constructor it needs to be able to create these gridelements
without an id. The Repository will assign an id as soon as the new
element gets persisted.