refactor moveelement
As preparation for ungrouping, we are now moving only references and not
elements. If no element exists at the target location, a new one is
created, and the old one is deleted. This leads to a more uniform
handling independent of if we are moving a single element, grouping,
ungrouping, or moving groups.