Feature/bc 495 ungrouping dashboard elements (#3126)
* adjust api

Can now pass groupIndex with a position to denote an element or position
within a group. Also fixes the api definition to export to openAPI
correctly.

* refactor gridelement to be created without id

For ungrouping elements we need to be able to create new gridelements
from the dashboard entity. Since it does not know about the repo, nor
the id constructor it needs to be able to create these gridelements
without an id. The Repository will assign an id as soon as the new
element gets persisted.

* refactor moveelement

As preparation for ungrouping, we are now moving only references and not
elements. If no element exists at the target location, a new one is
created, and the old one is deleted. This leads to a more uniform
handling independent of if we are moving a single element, grouping,
ungrouping, or moving groups.

* ungroup elements

moveElement can now move a single reference out of an element that is a
group.

* fix bug with moving the first reference of an element

an index of 0 was not triggering the special handling of a groupIndex, leading to a whole group being moved.

* extract getelement function