BC-507 gridreferences orphan removal (#3143)
* rename dashboard model mapper functions

mapToModel -> mapDashboardToModel
mapToEntity -> mapDashboardToEntity
these functions now follow the same naming convention as the other functions of the class.

* correctly orphan grid elements on persist

when a dashboard element is removed from a dashboard, its references are now correctly updated (removed)
This is done by explicitly removing each removed reference from the modelEntity.

* implement orphan removal

removed elements are now also fully deleted from the database.

* refactor dashboard model mapper to be injectable class

so far, the mapper was a static class, which not only made the code ugly
but also made testing unnecessarily difficult. It now is an injectable
nest service that is injected into the repo.

* add tests for model mapper

this is more because we can, to make future development on the file
easier by adding more tests as needed. The file already has complete
coverage through the integration tests of the repo.

* pass em as instance variable

the dashboard model mapper now recieves the entitymanager (em) as
instance variable at creation time. This enables us to remove it from
the function signatures.

* migration to reset dashboards collection

Since orphan removal now happens during writing, to ensure a smooth
transition we should ensure there are no orphans in the databases.
luckily, this is very easy, as we are only working with generated fake
data that can simply be dropped.

Co-authored-by: Martin Schuhmacher <55735359+MartinSchuhmacher@users.noreply.github.com>