Feature/bc 829 courses as dashboard elements (#3225)
* unidirectional reference to gridreferences

To support Courses in Grids, we first have to make sure the Course does
not need a reference back to the Grid. This change ensures that only the
gridElement needs to keep references of its referenced objects, but not vice
versa.

* change coursemetadata to support dashboard

both the frontend and the dashboard work with "title", while the course
works with "name". The coursemetadata interface has been changed to
return the name as "title" as well.

* dashboard elements now link to courses

courses now implement an interface known to dashboard
elements. (defined seperately in domain/interfaces)
to prepare for differently typed references, a "learnroomtype" has been
introduced
the dashboard model mapper now checks for the type of the references,
but only accepts courses.

* remove defaultGridReference

from domain and database model.
where necessary for tests, we are now using mocks or courses

* repo now generates empty dashboard

when a user has no dashboard, the dashboard returned by the repo will
now be a fresh, empty dashboard to avoid using fake objects

* add learnrooms to a dashboard entity

dashboard entity can now recieve a list of courses that will be added to
the dashboard.

* add newly created courses to dashboard in uc

whenever a user requests his dashboard, missing courses are added to
the dashboard.

* adjust default gridsize

grids can now be vertically endless to support an arbitrary number of
courses. Horizontally, the new size is for now 4, to work with all
frontend devices.

* migration to remove invalid dashboard data

* return courseIds in dashboard response

the ids returned in the gridelements will always be courseIds. The Ids
of the elements themselves, even for groups, are never given to the
frontend.