BC-1103 implementing lessons into rooms - Backend (#3298)
* implementing lessons into rooms - Backend

Additionally to tasks now lessons will be added to rooms view.
Teacher will see any lesson from the course while students
will only see lessons which are not hidden.

* implement tests for lessons repo

* adding board lesson to mapper

BoardLessonResponse was added as an alternative BoardElementResponse
with specific properties. BoardMapper was adapted to also support
the new BoardLessonResponse. With that it is also possible to easily
add new types of responses. Tests for all of that were added as well.

* adding LessonRepo to learnroom module for tests