BC-1178 - roomboard datastructure (#3316)
* implement board entities
Board and BoardElement have been implemented to serve as the datastructure for boards.

* implement syncing of lessons and tasks to board
new elements are added, old ones removed

* implement authorisation service for rooms
the board has to contain elements that not all users of the board are allowed to see.
The authorisation service implements logic to determine which elements have to be filtered.

* intermediary room dto
the current room entity structure does not completely fit the current
api dto structure. We thus have to add an intermediary dto class for now, to
map between domain and api layer. To avoid collisions, existing dtos
have been renamed to be more specific.
the intermediary dto is constructed by the uc, using a factory class.

* implement board repo
for fetching the board of a course, and for persisting changes.

* implement uc