collapse file single table inheritance
So far, the legacy files have used the single inheritance to savely
type files and directories. However, the discriminator value is a bool,
Which is not supported by mikroorm, which has lead to some issues.
In order to resolve these issues, the file entity has been reimplemented
without an inheritance structure.