ignore NOT_DOING when calculating final copy state
NOT_DOING elementes are things that are intentionally excluded from a copy.
They are thus not to be considered as failing. Instead, we ignore them
when calculating the final state of an element with multiple children.

This means other stati always take precedence, and an element with only
NOT_DOING children will fail
(as there are no children being considered.)