BC-4238 - column board feature toggle (#4185)
Found on staging: The Feature toggle for the column board does not work as expected. The automatic generation of a column-board instance for each course is triggered, although the feature is disabled for staging. This causes an error-page being displayed - due to the fact on staging (and probably on production) we have a race-condition when creating and reading the same object directly in sequence.