BC-7236 - refactor how to get rootid in socket gateway (#5080)
For any action that is executed on a board we have to find out who needs to get the resulting update. Socket.io organizes users in rooms. For the board - all clients on the same board are in the same room.
The room name is based on the boardid. So we always need to find out the rootId of the boardNode the action was executed on. For example: if a card title was rename, the card is the boardNode-object and has the rootId of the board the card is placed on.
*Until now*: the rootId was determined by (re)using the BoardAuthorisableService
*Now*: the rootId can be taken directly from the boardNode (due to boardPersisting-refactorings of Team MilkyWay) which reduces the number of queries that need to be executed

---------

Co-authored-by: hoeppner-dataport <marcus.hoeppner@dataport.de>