BC-8354 - adjust roles cache to use different tokens for different operations (#5322)
in the roles repo, findByNames and findByName used the same cache result (when only a single roleName was selected). This resulted in an error, when an array result was expected but the cache contained a single result, or when a single result was expected but cache contained an array with a single element.