BC-9010 - Runtime Configuration (#5903)
So far only a Module that can be used by other modules, without an API.
This means that values could only be edited in the database.

The Runtime Config is a dynamic module that can be used seperately by different Modules
and/or app. It is configured with default values per Instance, and ignores any values not
in its defaults.

* no deletion for now

When multiple Modules are using different instances of the Runtime
Config module, with different configurations of default values, they
would delete each other's values from the database.

Because of this, we will not automatically clean up old values from the
database for now.